backend/pyproject.toml
[project]
name = "hopperhacks25"
version = "0.1.0"
description = "Map for disabled people at stony that takes into account elevation + disability accesible routes"
authors = [
    {name = "tarunvaid05",email = "tarunvaid05@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "googlemaps (>=4.10.0,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "polyline (>=2.0.2,<3.0.0)",
    "networkx (>=3.4.2,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "scipy (>=1.15.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
